Upgrade gatsby-remark-relative-images to v2.0.2
Upgrade major version of gatsby-remark-relative-images from v0.3.0 to
v2.0.2.  Doing so requires some updates to the gatsby-node.js file to
remove a deprecated function.

Further, the dev server start process would fail due to a "Callback was
already called" error.  There doesn't seem to be a clear solution to the
issue, but what fixed it in the commit was:
* Switching to dart-sass from node-sass
* Use gatsby-plugin-netlify-cms v4.9.0 instead of using the latest
  version v4.10.0.

This is a workaround only.  References to track this issue are:
* https://github.com/gatsbyjs/gatsby/issues/24577
* https://github.com/gatsbyjs/gatsby/issues/27891
* https://github.com/gatsbyjs/gatsby/issues/25142